Move email sending to separate module
This makes it easier to test as a separate unit.
